pyproject.toml
[tool.poetry]
name = "qrproducts"
version = "0.1.0"
description = "Soft creado para conocer las características de cada producto, por medio del uso de qr."
authors = ["Carlos Valdemar Soria <carlosvsoria1982@gmail.com>"]

[tool.poetry.dependencies]
python = "3.8"
Flask = "2.0.2"
PyMySQL = "1.0.2"
python-dotenv = "^0.19.2"
cryptography = "^36.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
